% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACKAGE.R
\name{bootGeweke}
\alias{bootGeweke}
\title{bootGeweke: A Package for Bootstrap Based Geweke Diagnostics on MCMC Chains}
\description{
The \code{bootGeweke} package provides a function to perform a bootstrap-based Geweke diagnostic
on Markov Chain Monte Carlo (MCMC) objects. This package allows users to compute Z-scores
for Geweke diagnostics with bootstrap sampling, providing standard deviation and
confidence intervals for better interpretation of convergence diagnostics.
}
\section{Functions}{

\itemize{
\item \code{bootstrap_geweke()}: Perform a bootstrap-based Geweke diagnostic on an MCMC object.
}
}

