% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_geweke.R
\name{bootstrap_geweke}
\alias{bootstrap_geweke}
\title{Bootstrap Geweke Diagnostic}
\usage{
bootstrap_geweke(
  x,
  B = 1000,
  n = length(x),
  confidence_level = 0.95,
  frac1 = 0.1,
  frac2 = 0.5
)
}
\arguments{
\item{x}{An MCMC object, which can potentially contain more than one chain. If multiple chains are present, they will be combined into a single chain before the bootstrap process.}

\item{B}{Integer. The number of bootstrap samples to be generated. Default is 1000.}

\item{n}{Integer. The length of each bootstrap sample. By default, this is set to the same length as the combined chain.}

\item{confidence_level}{Numeric. The confidence level for calculating confidence intervals of Z-scores. Default is 0.95.}

\item{frac1}{Numeric. The proportion of the chain used for the initial segment. Default is 0.1 (the first 10\% of the chain).}

\item{frac2}{Numeric. The proportion of the chain used for the last segment. Default is 0.5 (the last 50\% of the chain).}
}
\value{
A list of class "bootGeweke" containing:
\describe{
\item{bootstrap_samples}{A matrix containing all bootstrap samples (B x n).}
\item{z_scores}{A vector of bootstrap Z-scores of length B.}
\item{z_mean}{The mean of the Z-scores.}
\item{z_sd}{The standard deviation of the Z-scores.}
\item{confidence_intervals}{A vector of confidence intervals for the Z-scores based on the provided confidence level.}
}
}
\description{
This function performs a bootstrap-based Geweke diagnostic on an MCMC object.
It computes Z-scores for the Geweke diagnostic, allowing for confidence intervals
based on the bootstrap samples.
}
\examples{
\dontrun{
mcmc_chain <- rnorm(1000)  # Example MCMC chain
result <- bootstrap_geweke(mcmc_chain, B = 500, n = 100, confidence_level = 0.95)
summary(result)
plot(result)
}

}
